ghc94: Switch PYTHON_FOR_BUILD_ONLY to "tool".
Setting it to "yes" performs a full buildlink3 inclusion of the necessary
python version.  This brings in gettext-lib, which on some platforms will add
an explicit -lintl to linker args, but gettext-lib is not marked as a full
dependency so this ends in runtime failures with missing libintl.

Switching to "tool" uses a plain DEPENDS instead so no problems.