font-util: update to 1.4.0.
This release updates the XORG_FONT_CHECK_COMPRESSION autoconf macro to
also set a COMPRESS_FLAGS variable.

By default, COMPRESS_FLAGS will be set to -n for gzip to stop recording
timestamps to improve build reproducibility, and to empty for all other
compression methods (compress, bzip2, or none).

This will not be effective until the bdf font packages we produce are
updated to add COMPRESS_FLAGS to Makefile.am, which will be occuring
shortly.

Alan Coopersmith (3):
      gitlab CI: stop requiring Signed-off-by in commits
      Add COMPRESS_FLAGS to pass options to compression command
      font-util 1.4.0